Update README.md
The Mahr in Islam: What It Is and Why It Matters

Mahr in Islam is a custom and an obligation. It is practiced in marriage ceremonies. It is basically, a portion of the money or can be any other worthy item that the groom is liable to his bride. Though there is not a definite amount or quantity of a product but it can be decided on a customary basis. The amount may differ based on a groom’s financial position and location. But it is a compulsion on the part of the groom to pay a certain amount to his bride.

Marriage is basically a legal contract in Islam and there are certain rules and regulation that is decided by the parties. There are a few obligatory practices and Mahr in Islam is one of them. A groom is obliged to pay Mahr and he cannot be exempted from it. Mahr fatimi is the amount that was provided to Hazrat Fatimah (Radhiyallah anha). Generally, two opinions regarding Mehr Fatimi exist.

The meaning of Mehr is the obligation that a male partner has to fulfill at the time of Islamic marriage to his female partner. When we translate Mehr in English, in simple terms it would be a gift or contribution that a husband offers to his bride. This is usually dubbed as financial independence for the bride or to provide her with a type of financial freedom. The ownership of Mahr goes to a female partner and no one can evade her from this right. 

Allah says in Quran: “And give the women (on marriage) their Mahr as a free gift” (Quran 4:4)

Surah An-Nisa, verse 20:

“But if you want to replace one wife with another and you have given one of them a great amount (in gifts), do not take back from it anything. Would you take it in injustice and manifest sin?” [4:20
 
 
Prophet PBUH said: “You are not entitled to take back any money. If you have told the truth, the Mahr you paid was for having sexual relations with her lawfully”. [Sahih Bukhari].

At another place, Prophet PBUH said: “Go and look for something, even if it is a ring of iron.” (Al-Bukhari)

Importance of Mahr.

Paying mahr is an obligation in Islamic marriage and is considered one of the main aspects. The woman is given Mehr as her exclusive property and to provide her with a sort of financial independence. Mehr in Islam is a liability and mandatory. It is paid because the girl usually holds nothing. So, to hold her rights, she is given Mehr. It is a sign of duty, respect, and responsibility. Further, it helps build a constructive relationship between husband and wife because it helps the husband understand the responsibilities and rights of his wife. 
  
FAQ:

what is the customary mahr in us as of today

Although the amount of Mehr in Islam is not defined and there is not a definite figure or number of properties or items. The amount may differ vary from country to country. Both husband and wife can decide the amount. If we take it in dollars then It could be millions of dollars or merely a few hundred dollars. But in customary terms, it is usually between 4 to 7 thousand US dollars in several Islamic countries.

How Much Mahr in Islam.

It is stated in Hadith that the Mahr amount should be one gold piece but it is not compulsory. Paying Mahr is a liability and obligation but the amount can be decided mutually by the guardians and there is not any restriction on the amount of mahr.

how to calculate mahr in Islam

Normally, guardians of both sides have to decide about the mahr amount. If they cannot then we can follow the normal pattern in that area. 


what is mahr in Islam
Mahr in Islam is the duty of the groom to pay sizeable money or property to his groom to provide her financial stability and independence. It is an obligation and clearly explained in Quran and Hadith and no one is exempted from it.